split out conversion of types
